Return Change
In this code , I have implemented a program that prompts the user the number of cents customer will get back and returns  the smallest number of coins with which that change can be made. If the user supplies invalid input then the code again asks for a valid input. Here I have used Greedy Algorithm.